requirements.txt
aiosqlite==0.20.0
asyncpg==0.29.0
bcrypt==4.2.0
cryptography==43.0.1
email-validator==2.2.0
environs==11.0.0
fastapi==0.115.0
loguru==0.7.2
PyJWT==2.9.0
python-multipart==0.0.9
SQLAlchemy==2.0.35
uvicorn==0.30.6
